# public/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: 'API macarte'
  version: '1.0'
servers:
  -
    url: 'https://macarte-qualif.ign.fr/'
    description: 'Accède aux données de l''environnement Ma Carte'
paths:
  /api/editorial/followers:
    get:
      tags:
        - Editorial
      operationId: e381179d562ca09a83d431b1dcfe77a5
      responses:
        '200':
          description: 'Liste des followers par réseau social'
          content:
            application/json:
              schema:
                properties:
                  facebook: { type: string, example: 65K }
                  twitter: { type: string, example: 13K }
                  linkedin: { type: string, example: 24K }
                  instagram: { type: string, example: 11K }
                type: object
        '404':
          $ref: '#/components/responses/NotFound'
  /api/editorial/megamenu:
    get:
      tags:
        - Editorial
      operationId: 4e6c574f04b8d8c4c2b65b6f58a53249
      responses:
        '200':
          description: 'Contenu html du megamenu'
          content:
            application/json:
              schema:
                properties:
                  html: { type: string }
                type: object
        '404':
          $ref: '#/components/responses/NotFound'
...
  '/api/editorial/articles/{category}':
    get:
      tags:
        - Editorial
      operationId: ba7f19a2655d46d5a131b6249ed0716b
      parameters:
        -
          name: category
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Liste des articles de la categorie'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Article'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/me:
    get:
      tags:
        - User
      operationId: 8c325367503eebff893de1e9a05f2efd
      responses:
        '200':
          description: 'Attributs de l''utilisateur connecté'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User_view'
        '401':
          $ref: '#/components/responses/NotConnected'
      security:
        -
          scalar: bearer
    patch:
      tags:
        - User
      description: "parameter **current_password** is required to patch username, email and new_password<br>\r\n                username, email et public_name sont uniques, la réponse 400 '... already used' peut être renvoyée\r\n     "
      operationId: 565897838a26198f5f59e59240e4e354
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User_me_edit'
      responses:
        '200':
          description: 'Attributs de l''utilisateur'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User_view'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NotConnected'
      security:
        -
          scalar: bearer
...
  /api/login:
    post:
      tags:
        - Login
      operationId: c19de2fbeb914a1eb095ea4bc8242d9b
      parameters:
        -
          name: username
          in: query
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Jeton Jwt et rafraichissement'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Login'
        '401':
          description: 'Erreur dans le nom d''utilisateur ou le mot de passe'
          content:
            application/json:
              schema:
                properties:
                  code: { type: integer, example: '401' }
                  message: { type: string, example: 'Bad credentials' }
                type: object
        '429':
          description: 'Trop de tentatives de connexion échouées'
          content:
            application/json:
              schema:
                properties:
                  code: { type: integer, example: '429' }
                  message: { type: string, example: '5 tentatives de connexion échouées, allez sur /debloquer-mon-compte' }
                type: object
  /api/token/refresh:
    post:
      tags:
        - Login
      operationId: 26a9d1dfbaf768f672b4d0eba7d2a663
      parameters:
        -
          name: refresh_token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Jeton JWT et refresh'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Login'
        '401':
          description: 'An authentication exception occurred'
          content:
            application/json:
              schema:
                properties:
                  code: { type: integer, example: '401' }
                  message: { type: string, example: 'An authentication exception occurred' }
                type: object
  /api/logout:
    post:
      tags:
        - Login
      operationId: 17cd40d6062e1990f0e90edddd985c80
      responses:
        '200':
          description: logout
        '401':
          $ref: '#/components/responses/NotConnected'
      security:
        -
          scalar: bearer
...
components:
  schemas:
    Article:
      properties:
        id:
          description: 'Identifiant de l''article'
          type: integer
        category:
          description: 'Categorie de l''article'
          type: string
        position:
          description: 'Ordre d''affichage de l''article'
          type: number
        title:
          description: 'Titre de l''article'
          type: string
        content:
          description: 'Contenu de l''article'
          type: string
        tags:
          description: 'Tags de l''article'
          type: array
          items:
            type: string
        updated_at:
          description: 'Date de mise à jour de l''article'
          type: string
          format: date-time
        updated_by:
          description: 'Nom public de l''utilisateur qui a modifié l''article'
          type: string
        img_url:
          description: 'Url de l''image d''illustration de l''article'
          type: string
          format: uri
        link_text:
          description: 'Article de la catégorie news : texte du lien'
          type: string
        link_url:
          description: 'Article de la catégorie news : destination du lien'
          type: string
          format: uri
      type: object
...
    Login:
      properties:
        token:
          description: 'Jeton JWT à envoyer dans le header Authorization: Bearer xxx'
          type: string
        refresh_token:
          description: 'Jeton pour rafraichir le jeton JWT à envoyer /api/token/refresh'
          type: string
      type: object
    User_public:
      properties:
        public_id:
          description: 'identifiant public de l''utilisateur, ne sera jamais modifié'
          type: string
        public_name:
          description: 'Unique, ne doit pas être déjà utilisé (Erreur 400)'
          type: string
        twitter_account:
          description: 'Nom de compte Twitter'
          type: string
        facebook_account:
          description: 'Nom de compte Facebook'
          type: string
        linkedin_account:
          description: 'Nom de compte LinkedIn'
          type: string
        presentation:
          description: 'Présentation publique'
          type: string
        profile_picture:
          description: 'Url de l''image de profil'
          type: string
          format: uri
        cover_picture:
          description: 'Url de l''image de couverture'
          type: string
          format: uri
      type: object
    User_edit:
      type: object
      allOf:
        -
          properties:
            email:
              description: 'Unique, ne doit pas être déjà utilisé (Erreur 400)'
              type: string
              format: email
            username:
              description: 'Unique, ne doit pas être déjà utilisé (Erreur 400)'
              type: string
        -
          $ref: '#/components/schemas/User_public'
    User_me_edit:
      type: object
      allOf:
        -
          properties:
            current_password:
              type: string
            new_password:
              type: string
        -
          $ref: '#/components/schemas/User_edit'
    User_view:
      type: object
      allOf:
        -
          properties:
            locked:
              description: 'Compte bloqué'
              type: boolean
            id:
              type: integer
            medias_limit_size:
              description: 'Taille globale limite des images uploadées (en octet)'
              type: integer
            medias_size:
              description: 'Taille globale des images uploadées (en octet)'
              type: integer
            roles:
              type: array
              items:
                type: string
        -
          $ref: '#/components/schemas/User_edit'
  responses:
...
    BadRequest:
      description: 'La requête n''est pas correcte'
      content:
        application/json:
          schema:
            properties:
              code:
                type: integer
                example: '404'
              message:
                type: string
                example: 'La requête n''est pas correcte'
            type: object
    NotConnected:
      description: 'Vous devez être connecté'
      content:
        application/json:
          schema:
            properties:
              code:
                type: integer
                example: '401'
              message:
                type: string
                example: 'Vous devez être connecté'
            type: object
    NotFound:
      description: 'La resource n''existe pas'
      content:
        application/json:
          schema:
            properties:
              code:
                type: integer
                example: '404'
              message:
                type: string
                example: 'La resource n''existe pas'
            type: object
...
  securitySchemes:
    bearer:
      type: apiKey
      name: Authorization
      in: header
      bearerFormat: JWT
      scheme: Bearer